fix: cinematic state persistence, dedup start/stop, easing floor, size_t cast
- Persist cinematic panel visibility as member instead of local
- Extract StartCinematic/StopCinematic to eliminate 3x duplication
- Add 10% minimum speed floor to easing curve to prevent stalling
- Use size_t for pixel buffer allocation to prevent overflow

+++ src/render/cinematic/CaptureManager.cpp
@@ -52,7 +52,7 @@ bool CaptureManager::CaptureScreenshot(int width, int height)
     if (!EnsureCaptureDirectory())
         return false;
 
-    _pixelBuffer.resize(width * height * CaptureDefaults::RgbChannels);
+    _pixelBuffer.resize(static_cast<size_t>(width) * height * CaptureDefaults::RgbChannels);
     glReadPixels(0, 0, width, height, GL_RGB, GL_UNSIGNED_BYTE, _pixelBuffer.data());
 
     // Flip vertically (OpenGL origin is bottom-left, PNG expects top-left)
